Adding lots of changes after trip

+++ frontend/src/components/Dashboard.js
@@ -1,26 +1,43 @@
-import React from 'react';
+// import React from 'react';
+import React, { useState,useEffect } from 'react';
 import NavbarComponent from './includes/Navbar';
 import { useSelector } from 'react-redux';
-import { Navigate,useNavigate } from 'react-router-dom';
+import { 
+  Navigate,
+  // useNavigate 
+} from 'react-router-dom';
 import { useDispatch } from 'react-redux';
-import { deleteUser } from '../features/user';
+import { deleteUser,logout } from '../features/user';
+import M from 'materialize-css';
 
 function Dashboard() {
     const { isAuthenticated, user, loading } = useSelector(state => state.user);
-    const navigate = useNavigate();
     const dispatch = useDispatch();
+    const [confirmDelete, setConfirmDelete] = useState(false);
+
+    useEffect(() => {
+      // Initialize Materialize CSS toast
+      M.AutoInit();
+    }, []);
 
     const handleDeleteUser = () => {
-      dispatch(deleteUser(user.id))
-      .unwrap()
-      .then(() => {
-        // Deletion was successful, handle any necessary actions
-        console.log('User deleted successfully');
-      })
-      .catch((error) => {
-        // Error occurred during deletion, handle the error
-        console.error('Error deleting user:', error);
-      });
+      if (confirmDelete) {
+        M.toast({ html: 'User deleted successfully' });
+        dispatch(deleteUser(user.id))
+          .unwrap()
+          .then(() => {
+            console.log('User deleted successfully');
+            
+            // Redirect to the homepage
+            // return <Navigate to='/' />;
+          })
+          .catch((error) => {
+            console.log('Error deleting user:', error);
+          });
+          dispatch(logout());
+      } else {
+        setConfirmDelete(true);
+      }
     };
 
     if (!isAuthenticated && !loading && user === null)
@@ -39,9 +56,9 @@ function Dashboard() {
               <div className='card'>
                 <div className='card-body'>
                   <h5 className='card-title center'>User Details</h5>
-                  <div className='row'>
-                    <div className='col-12'>Profile Picture:</div>
-                    <div className='col'>
+                  {/* <div className='row'> */}
+                    {/* <div className='col-12'>Profile Picture:</div> */}
+                    {/* <div className='col'>
                       {user.profile_picture ? (
                         <img
                           src={user.profile_picture}
@@ -52,8 +69,8 @@ function Dashboard() {
                       ) : (
                         'No picture uploaded'
                       )}
-                    </div>
-                  </div>
+                    </div> */}
+                  {/* </div> */}
                   <div className='row'>
                     <div className='col-2'>Username:</div>
                     <div className='col'>{user.username}</div>
@@ -62,24 +79,39 @@ function Dashboard() {
                     <div className='col-2'>Email:</div>
                     <div className='col'>{user.email}</div>
                   </div>
-                  <div className='row'>
+                  {/* <div className='row'>
                     <div className='col-2'>First Name:</div>
                     <div className='col'>{user.first_name}</div>
                   </div>
                   <div className='row'>
                     <div className='col-2'>Last Name:</div>
                     <div className='col'>{user.last_name}</div>
-                  </div>
+                  </div> */}
                 </div>
               </div>
-              <div>
+              {/* <div>
                 <button className='btn btn-primary mt-3' onClick={() => navigate('/update')}>
                   Update Profile
                 </button>
-              </div>
-              <div>
+              </div> */}
+              {/* <div>
                 <button className='btn btn-primary mt-3' onClick={handleDeleteUser}>Delete User</button>
-              </div>
+              </div> */}
+              {!confirmDelete ? (
+              <button className='btn btn-primary mt-3' onClick={() => setConfirmDelete(true)}>
+                Delete User
+              </button>
+            ) : (
+              <>
+                <p>Are you sure you want to delete this user? All your requests will be lost forever.</p>
+                <button className='btn btn-danger mt-3' onClick={handleDeleteUser}>
+                  Yes
+                </button>
+                <button className='btn btn-secondary mt-3' onClick={() => setConfirmDelete(false)}>
+                  No
+                </button>
+              </>
+            )}
             </div>
           )}
         </>
